requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
python-jose[cryptography]
passlib[bcrypt]
tensorflow 
tensorflow-text 
tensorflow-hub
gradio_client
jinja2
gradio
Pillow